Move TestCountingSource to appropriate location
